new input.js to handle keyboard/mouse/scrolls
